Move connector-manager to the correct location
