+++ lib/dialects/abstract/connector-manager.js
