+++ lib/dialects/mariadb/connector-manager.js
@@ -112,7 +112,7 @@ module.exports = (function() {
               if (connection) {
                 connection.queryType = 'write'
               }
-              
+
               done(err, connection)
             }, self.config.replication.write)
           },
@@ -152,11 +152,11 @@ module.exports = (function() {
 
       return
     }.bind(this);
-    
+
     process.on('exit', this.onProcessExit)
   }
 
-  Utils._.extend(ConnectorManager.prototype, require("../connector-manager").prototype)
+  Utils._.extend(ConnectorManager.prototype, require("../abstract/connector-manager").prototype)
 
   ConnectorManager.prototype.query = function(sql, callee, options) {
     if (!this.isConnected && !this.pool) {
@@ -287,7 +287,7 @@ module.exports = (function() {
     client
       .on('error', function (err) {
         self.isConnecting = false
-       
+
         done(err)
       })
       .on('connect', function () {
