+++ lib/dialects/mysql/connector-manager.js
@@ -153,11 +153,11 @@ module.exports = (function() {
 
       return
     }.bind(this);
-    
+
     process.on('exit', this.onProcessExit)
   }
 
-  Utils._.extend(ConnectorManager.prototype, require("../connector-manager").prototype);
+  Utils._.extend(ConnectorManager.prototype, require("../abstract/connector-manager").prototype);
 
   ConnectorManager.prototype.query = function(sql, callee, options) {
     if (this.useQueue) {
@@ -185,7 +185,7 @@ module.exports = (function() {
           setTimeout(function() {
             if (self.pendingQueries === 0){
               self.disconnect.call(self);
-            } 
+            }
           }, 100);
         }
       }
