+++ lib/dialects/postgres/connector-manager.js
@@ -27,7 +27,7 @@ module.exports = (function() {
 
     process.on('exit', this.onProcessExit)
   }
-  Utils._.extend(ConnectorManager.prototype, require("../connector-manager").prototype)
+  Utils._.extend(ConnectorManager.prototype, require("../abstract/connector-manager").prototype)
 
   ConnectorManager.prototype.endQuery = function() {
     var self = this
