+++ lib/dialects/sqlite/connector-manager.js
@@ -14,7 +14,7 @@ module.exports = (function() {
     }
   }
 
-  Utils._.extend(ConnectorManager.prototype, require("../connector-manager").prototype)
+  Utils._.extend(ConnectorManager.prototype, require("../abstract/connector-manager").prototype)
 
   ConnectorManager.prototype.connect = function() {
     var emitter = new (require('events').EventEmitter)()
